+++ src/features/identityCheck/pages/identification/errors/eduConnect/EduConnectErrorBoundary.tsx
@@ -1,7 +1,7 @@
 import React, { ComponentType, memo, useEffect } from 'react'
 import { withErrorBoundary, FallbackProps } from 'react-error-boundary'
 
-import { NotEligibleEduConnect } from 'features/identityCheck/errors/eduConnect/NotEligibleEduConnect'
+import { NotEligibleEduConnect } from 'features/identityCheck/pages/identification/errors/eduConnect/NotEligibleEduConnect'
 import { eventMonitoring } from 'libs/monitoring'
 
 import { EduConnectError } from './types'
